Support class node serialization
Since the JSR-223 does not define a code sharing mechanism, the only option for
emulating this is by adding multiple source units into a script's compilation unit
before compiling it. However when a dependency is modified, all scripts referencing
it need to be recompiled as well, which is far from optimal. Therefore, an option
could be injecting serializable class nodes into a script's AST prior to its
compilation.